requirements.txt
cassandra-driver==3.11.0
Flask==0.12.2
Flask-RESTful==0.3.6